Use Jekyll 2.4.0 and fix url structure
Ok so. Github pages does not use the current version of jekyll which
results in the nice permalink default settings that are available on
2.5.3 not work.

So instead of doing

```
defaults:
  -
    scope:
      ...
    values:
      permalinks: "/posts/:title"
```

we need to set it outside and do
```
permalink: "/posts/:title"
```

which for some reason only sets it for the posts, but I'm guessing
that's because the pages have explict permalinks set.